+++ include/MobilityInterface/MobilityInterface.h
@@ -4,7 +4,7 @@
  */
 #ifndef MOBILITYINTERFACE_H
 #define MOBILITYINTERFACE_H
-#include "container.h"
+#include "memory/container.h"
 #include "defines.h"
 #include "lanczos.h"
 #include <random>
